create omnibox module
