+++ webpack.config.js
@@ -11,10 +11,7 @@ module.exports = {
       {
         test: /\.js$/,
         exclude: /(node_modules|bower_components)/,
-        loader: "babel-loader",
-        query: {
-          presets: ["env"]
-        }
+        loader: "babel-loader"
       }
     ]
   }
